added ability for admins to view deleted songs, and updated activity for when deleted

+++ src/Views/songs/index-manage-buttons.blade.php
@@ -1,11 +1,21 @@
-{{ link_to_route('songs.show', 'View', ['id' => $song->id], ['class' => 'btn btn-info btn-xs']) }}
+@if(!$song->trashed())
+    {{ link_to_route('songs.show', 'View', ['id' => $song->id], ['class' => 'btn btn-info btn-xs']) }}
 
-@if($currentUser->is_admin == 'y' || $song->votes()->count() <= 1)
-    {{ link_to_route('songs.edit', 'Edit', ['id' => $song->id], ['class' => 'btn btn-primary btn-xs']) }}
+    @if($currentUser->is_admin == 'y' || $song->votes()->count() <= 1)
+        {{ link_to_route('songs.edit', 'Edit', ['id' => $song->id], ['class' => 'btn btn-primary btn-xs']) }}
+    @endif
 @endif
 
 @if($currentUser->is_admin == 'y' || $currentUser->id == $song->user_id)
-    {{ Form::open(['route' => ['songs.destroy', $song->id], 'method' => 'delete', 'style' => 'display:inline;', 'data-confirm' => 'Are you sure?', 'class' => 'delete-confirm']) }}
-    <button type="submit" class="btn btn-danger btn-xs">Delete</button>
-    {{ Form::close() }}
+    @if($song->trashed())
+        {{ link_to_route('songs.edit', 'Restore', ['id' => $song->id], ['class' => 'btn btn-success btn-xs']) }}
+
+        {{ Form::open(['route' => ['songs.destroy', $song->id], 'method' => 'delete', 'style' => 'display:inline;', 'data-confirm' => 'Are you sure?', 'class' => 'delete-confirm']) }}
+        <button type="submit" class="btn btn-danger btn-xs">Force Delete</button>
+        {{ Form::close() }}
+    @else
+        {{ Form::open(['route' => ['songs.destroy', $song->id], 'method' => 'delete', 'style' => 'display:inline;', 'data-confirm' => 'Are you sure?', 'class' => 'delete-confirm']) }}
+        <button type="submit" class="btn btn-danger btn-xs">Delete</button>
+        {{ Form::close() }}
+    @endif
 @endif
